Filtered the model list to remove duplicate IDs
temporary fix; The names and other properties could be different, but the requests are routed purely on the basis of the model ID, which had duplicate IDs for variants. 
A permanent fix is called for to have unique IDs for all variants.

Filtering also provides a better experience for the demo by removing redundancy.